+++ src/routers/MsTeamsPageRouter.ts
@@ -3,6 +3,7 @@
 
 import { Router } from "express";
 import * as path from "path";
+import * as debug from "debug";
 
 /**
  * Page Router options
@@ -14,18 +15,18 @@ export interface IMsTeamsPageRouterOptions {
     root: string;
 
     /**
-     * All components
+     * All components, used to detect presence of `PreventIframe` files
      */
     components: any;
 }
 
 /**
  * Express router for pages to be hosted in Microsoft Teams.
- * Pages MUST have extension of html and end with Tab, Config, Remove, Connector or ConnectorConnected
  * @param options Page Router options
  */
 export const MsTeamsPageRouter = (options: IMsTeamsPageRouterOptions): Router => {
     const router = Router();
+    const log = debug("msteams");
 
     // This is used to prevent your tabs from being embedded in other systems than Microsoft Teams
     router.use((req: any, res: any, next: any) => {
@@ -39,13 +40,16 @@ export const MsTeamsPageRouter = (options: IMsTeamsPageRouterOptions): Router =>
     for (const app in options.components) {
         if (options.components.hasOwnProperty(app)) {
             const component = options.components[app];
-            if (component.__addCsp__) {
-                const arr: string[] = component.__addCsp__;
-                arr.forEach((page) => {
-                    router.get(page, (req: any, res: any, next: any) => {
-                        res.sendFile(path.join(options.root, req.path));
+            if (component.__addCsp) {
+                const arr: string[] = component.__addCsp;
+                if (arr.length) {
+                    arr.forEach((page) => {
+                        log(`Adding CSP policy for ${page}`);
+                        router.get(page, (req: any, res: any, next: any) => {
+                            res.sendFile(path.join(options.root, req.path));
+                        });
                     });
-                });
+                }
             }
         }
     }